backend/pyproject.toml
[project]
name = "microfish-backend"
version = "1.0.0"
description = "Microfish - Cloud-powered multi-agent simulation engine with Azure OpenAI + GraphRAG"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Anup Roy" }
]

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "numpy>=1.24.0",
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    "PyMuPDF>=1.24.0",
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
